requirements.txt
Flask==2.2.2
Flask_Login==0.6.2
Flask_Migrate==3.1.0
Flask_SQLAlchemy==2.5.1
PyJWT==2.5.0
pymongo==4.2.0
python-dotenv==0.21.0
bcrypt==4.0.0
psycopg2
